backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg>=0.30.0
pgvector==0.2.4
openai==1.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
newspaper3k==0.2.8
feedparser==6.0.11
httpx==0.26.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
tavily-python==0.3.0
lxml_html_clean==0.4.3
slowapi==0.1.9
